// src/api/fiberlaser/models/PaginatedListPlatesResponseDtoDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { ListPlatesResponseDto } from "./ListPlatesResponseDto";

export type PaginatedListPlatesResponseDtoDto = {
    /**
     * @type number
    */
    total: number;
    /**
     * @type number
    */
    page: number;
    /**
     * @type number
    */
    limit: number;
    /**
     * @type number
    */
    totalPages: number;
    /**
     * @description Lista de itens paginados do tipo ListPlatesResponseDto.
     * @type array
    */
    data: ListPlatesResponseDto[];
};